Added support for IPC communication between rendered process and electron process, removing the need for an exposed API.
