+++ api/routes/auth.js
@@ -1,5 +1,7 @@
-const url = require('url');
 const express = require('express');
+const {
+  authHandler,
+} = require('../../electron/auth');
 
 const router = express.Router();
 
@@ -7,76 +9,7 @@ module.exports = (app, auth) => {
   router.post('/', auth.authenticate('saml', {
     failureFlash: true,
     failureRedirect: app.get('configureUrl'),
-  }), (req, res) => {
-    let roleAttr = req.user['https://aws.amazon.com/SAML/Attributes/Role'];
-    let frontend = process.env.ELECTRON_START_URL || app.get('baseUrl');
-
-    frontend = new url.URL(frontend);
-
-    // Convert roleAttr to an array if it isn't already one
-    if (!Array.isArray(roleAttr)) {
-      roleAttr = [roleAttr];
-    }
-
-    const roles = roleAttr.map((arns, i) => {
-      const [roleArn, principalArn] = arns.split(',');
-      const roleArnSegments = roleArn.split(':');
-      const accountId = roleArnSegments[4];
-      const roleName = roleArnSegments[5].replace('role/', '');
-
-      return {
-        accountId,
-        index: i,
-        principalArn,
-        roleArn,
-        roleName,
-      };
-    });
-
-    const session = req.session.passport;
-
-    session.samlResponse = req.body.SAMLResponse;
-    session.roles = roles;
-
-    if (roles.length > 1) {
-      // If the session has a previous role, see if it matches
-      // the latest roles from the current SAML assertion.  If it
-      // doesn't match, wipe it from the session.
-      if (session.roleArn && session.principalArn) {
-        const found = roles
-          // eslint-disable-next-line max-len
-          .find((role) => role.roleArn === session.roleArn && role.principalArn === session.principalArn);
-
-        if (!found) {
-          session.showRole = undefined;
-          session.roleArn = undefined;
-          session.roleName = undefined;
-          session.principalArn = undefined;
-          session.accountId = undefined;
-        }
-      }
-
-      // If the session still has a previous role, proceed directly to auth.
-      // Otherwise ask the user to select a role.
-      if (session.roleArn && session.principalArn && session.roleName && session.accountId) {
-        frontend.searchParams.set('auth', 'true');
-      } else {
-        frontend.searchParams.set('select-role', 'true');
-      }
-    } else {
-      const role = roles[0];
-
-      frontend.searchParams.set('auth', 'true');
-
-      session.showRole = false;
-      session.roleArn = role.roleArn;
-      session.roleName = role.roleName;
-      session.principalArn = role.principalArn;
-      session.accountId = role.accountId;
-    }
-
-    res.redirect(frontend);
-  });
+  }), authHandler(app));
 
   return router;
 };
