--- api/routes/logout.js
@@ -1,15 +1,0 @@
-const express = require('express');
-
-const router = express.Router();
-
-module.exports = (app) => {
-  router.get('/', (req, res) => {
-    app.set('entryPointUrl', null);
-    req.session.destroy();
-    res.json({
-      logout: true,
-    });
-  });
-
-  return router;
-};
