+++ electron/logout.js
@@ -1,0 +1,16 @@
+const {
+  app,
+} = require('../api/server');
+
+async function logout() {
+  Storage.set('session', {});
+  app.set('entryPointUrl', null);
+
+  return {
+    logout: true,
+  };
+}
+
+module.exports = {
+  logout,
+};
