+++ package.json
@@ -58,7 +58,6 @@
     "@fortawesome/free-solid-svg-icons": "^6.3.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@popperjs/core": "^2.11.6",
-    "axios": "^1.3.4",
     "babel-jest": "^29.5.0",
     "bootstrap": "^5.2.3",
     "coveralls": "^3.1.1",
